separate filter request test
